requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
boto3==1.28.7
botocore==1.31.7
certifi==2023.5.7
charset-normalizer==3.2.0
cmake==3.27.0
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.3
executing==1.2.0
filelock==3.12.2
fonttools==4.41.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
s3transfer==0.6.1
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.138
urllib3==1.26.16
wcwidth==0.2.6
pandas
